automation/requirements.txt
# Fábrica de Liberdade — Automação de Conteúdo
# pip install -r requirements.txt

anthropic>=0.30.0
openai>=1.30.0
firebase-admin>=6.5.0
requests>=2.31.0
python-slugify>=8.0.3
python-dotenv>=1.0.1
pyairtable>=2.3.0
schedule>=1.2.1
Pillow>=10.3.0
httpx>=0.27.0
google-auth>=2.29.0
pytrends>=4.9.2
beautifulsoup4>=4.12.3
lxml>=5.2.1
